requirements.txt
flask
chromadb
openai
python-dotenv
PyPDF2